+++ bin/cmds/docs.d.ts
@@ -1,0 +1,1 @@
+export declare const docs: (folders: string) => void;
