+++ bin/index.d.ts
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+export {};
