+++ bin/util/table.d.ts
