+++ bin/util/util.d.ts
