--- lib/index.d.ts
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-declare const chalk: any;
-declare const clear: any;
-declare const figlet: any;
-declare const path: any;
-declare const program: any;
