--- lib/index.js
@@ -1,25 +1,0 @@
-#!/usr/bin/env node
-"use strict";
-var chalk = require('chalk');
-var clear = require('clear');
-var figlet = require('figlet');
-var path = require('path');
-var program = require('commander');
-clear();
-console.log(chalk.red(figlet.textSync('Directus Toolkit', { horizontalLayout: 'full' })));
-program
-    .version('0.0.1')
-    .description("Do some cool stuff and save some time developing with directus!")
-    .option('-p, --peppers', 'Add peppers')
-    .option('-P, --pineapple', 'Add pineapple')
-    .option('-b, --bbq', 'Add bbq sauce')
-    .option('-c, --cheese <type>', 'Add the specified type of cheese [marble]')
-    .option('-C, --no-cheese', 'You do not want any cheese')
-    .parse(process.argv);
-console.log('you ordered a pizza with:');
-if (program.peppers)
-    console.log('  - peppers');
-if (program.pineapple)
-    console.log('  - pineapple');
-if (program.bbq)
-    console.log('  - bbq');
