--- lib/templates/interface/interface.story.d.ts
@@ -1,3 +1,0 @@
-declare const _default: {};
-export default _default;
-export declare const basic: () => any;
