--- lib/templates/interface/interface.test.d.ts
@@ -1,1 +1,0 @@
-export {};
