+++ src/templates/interface/interface.story.txt
